requirements.txt
python-telegram-bot
python-dotenv
google-generativeai
yfinance==0.2.61
pandas